[system] Improve the SxProp typings structure, by splitting them in a separate module. (#24090)
