+++ packages/material-ui-system/src/index.d.ts
@@ -1,5 +1,6 @@
 import * as React from 'react';
 import * as CSS from 'csstype';
+import { CSSProperties } from './CSSProperties';
 // disable automatic export
 export {};
 
@@ -262,29 +263,12 @@ export type TypographyProps = PropsFor<typeof typography>;
 // eslint-disable-next-line @typescript-eslint/naming-convention
 export function unstable_getThemeValue(prop: string, value: any, theme: object): any;
 
-export type StandardCSSProperties = CSS.PropertiesFallback<number | string>;
-
 /**
  * The `css` function accepts arrays as values for mobile-first responsive styles.
  * Note that this extends to non-theme values also. For example `display=['none', 'block']`
  * will also works.
  */
 export type ResponsiveStyleValue<T> = T | Array<T | null> | { [key: string]: T | null };
-
-/**
- * All non-vendor-prefixed CSS properties. (Also allows `number` in order to support CSS-in-JS libs,
- * since they are converted to `px`.)
- */
-export interface CSSProperties
-  extends CSS.StandardProperties<number | string>,
-    CSS.SvgProperties<number | string> {}
-
-/**
- * Map of all CSS pseudo selectors (`:hover`, `:focus`, ...).
- */
-export type CSSPseudoSelectorProps<Theme extends object = {}> = {
-  [K in CSS.Pseudos]?: SystemStyleObject<Theme>;
-};
 
 /**
  * CSS as a plain object that is compatible with CSS-in-JS libraries.
@@ -304,368 +288,5 @@ export interface CSSOthersObjectForCSSObject {
   [propertiesName: string]: CSSInterpolation;
 }
 
-/**
- * Map all nested selectors.
- */
-export interface CSSSelectorObject<Theme extends object = {}> {
-  [cssSelector: string]: SystemStyleObject<Theme>;
-}
-
-export interface AliasesCSSProperties {
-  /**
-   * The **`background-color`** CSS property sets the background color of an element.
-   *
-   * **Initial value**: `transparent`
-   *
-   * | Chrome | Firefox | Safari |  Edge  |  IE   |
-   * | :----: | :-----: | :----: | :----: | :---: |
-   * | **1**  |  **1**  | **1**  | **12** | **4** |
-   *
-   * @see https://developer.mozilla.org/docs/Web/CSS/background-color
-   */
-  bgcolor?: StandardCSSProperties['backgroundColor'];
-  /**
-   * The **`margin`** CSS property sets the margin on all four sides of an element. It is a shorthand for `margin-top`, `margin-right`, `margin-bottom`, and `margin-left`.
-   *
-   * | Chrome | Firefox | Safari |  Edge  |  IE   |
-   * | :----: | :-----: | :----: | :----: | :---: |
-   * | **1**  |  **1**  | **1**  | **12** | **3** |
-   *
-   * @see https://developer.mozilla.org/docs/Web/CSS/margin
-   */
-  m?: StandardCSSProperties['margin'];
-  /**
-   * The **`margin-top`** CSS property sets the margin on the top of an element. A positive value places it farther from its neighbors, while a negative value places it closer.
-   *
-   * **Initial value**: `0`
-   *
-   * | Chrome | Firefox | Safari |  Edge  |  IE   |
-   * | :----: | :-----: | :----: | :----: | :---: |
-   * | **1**  |  **1**  | **1**  | **12** | **3** |
-   *
-   * @see https://developer.mozilla.org/docs/Web/CSS/margin-top
-   */
-  mt?: StandardCSSProperties['marginTop'];
-  /**
-   * The **`margin-right`** CSS property sets the margin on the right side of an element. A positive value places it farther from its neighbors, while a negative value places it closer.
-   *
-   * **Initial value**: `0`
-   *
-   * | Chrome | Firefox | Safari |  Edge  |  IE   |
-   * | :----: | :-----: | :----: | :----: | :---: |
-   * | **1**  |  **1**  | **1**  | **12** | **3** |
-   *
-   * @see https://developer.mozilla.org/docs/Web/CSS/margin-right
-   */
-  mr?: StandardCSSProperties['marginRight'];
-  /**
-   * The **`margin-bottom`** CSS property sets the margin on the bottom of an element. A positive value places it farther from its neighbors, while a negative value places it closer.
-   *
-   * **Initial value**: `0`
-   *
-   * | Chrome | Firefox | Safari |  Edge  |  IE   |
-   * | :----: | :-----: | :----: | :----: | :---: |
-   * | **1**  |  **1**  | **1**  | **12** | **3** |
-   *
-   * @see https://developer.mozilla.org/docs/Web/CSS/margin-bottom
-   */
-  mb?: StandardCSSProperties['marginBottom'];
-  /**
-   * The **`margin-left`** CSS property sets the margin on the left side of an element. A positive value places it farther from its neighbors, while a negative value places it closer.
-   *
-   * **Initial value**: `0`
-   *
-   * | Chrome | Firefox | Safari |  Edge  |  IE   |
-   * | :----: | :-----: | :----: | :----: | :---: |
-   * | **1**  |  **1**  | **1**  | **12** | **3** |
-   *
-   * @see https://developer.mozilla.org/docs/Web/CSS/margin-left
-   */
-  ml?: StandardCSSProperties['marginLeft'];
-  /**
-   * The **`mx`** property is shorthand for using both **`margin-left`** and **`margin-right`** CSS properties. They set the margin on the left and right side of an element. A positive value places it
-   * farther from its neighbors, while a negative value places it closer.
-   *
-   * **Initial value**: `0`
-   *
-   * | Chrome | Firefox | Safari |  Edge  |  IE   |
-   * | :----: | :-----: | :----: | :----: | :---: |
-   * | **1**  |  **1**  | **1**  | **12** | **3** |
-   *
-   * @see https://developer.mozilla.org/docs/Web/CSS/margin-left
-   * @see https://developer.mozilla.org/docs/Web/CSS/margin-right
-   */
-  mx?: StandardCSSProperties['marginLeft'];
-  /**
-   * The **`marginX`** property is shorthand for using both **`margin-left`** and **`margin-right`** CSS properties. They set the margin on the left and right side of an element. A positive value
-   * places it farther from its neighbors, while a negative value places it closer.
-   *
-   * **Initial value**: `0`
-   *
-   * | Chrome | Firefox | Safari |  Edge  |  IE   |
-   * | :----: | :-----: | :----: | :----: | :---: |
-   * | **1**  |  **1**  | **1**  | **12** | **3** |
-   *
-   * @see https://developer.mozilla.org/docs/Web/CSS/margin-left
-   * @see https://developer.mozilla.org/docs/Web/CSS/margin-right
-   */
-  marginX?: StandardCSSProperties['marginLeft'];
-  /**
-   * The **`my`** property is shorthand for using both **`margin-top`** and **`margin-bottom`** CSS properties. They set the margin on the top and bottom of an element. A positive value places it
-   * farther from its neighbors, while a negative value places it closer.
-   *
-   * **Initial value**: `0`
-   *
-   * | Chrome | Firefox | Safari |  Edge  |  IE   |
-   * | :----: | :-----: | :----: | :----: | :---: |
-   * | **1**  |  **1**  | **1**  | **12** | **3** |
-   *
-   * @see https://developer.mozilla.org/docs/Web/CSS/margin-top
-   * @see https://developer.mozilla.org/docs/Web/CSS/margin-bottom
-   */
-  my?: StandardCSSProperties['marginTop'];
-  /**
-   * The **`marginY`** property is shorthand for using both **`margin-top`** and **`margin-bottom`** CSS properties. They set the margin on the top and bottom of an element. A positive value places
-   * it farther from its neighbors, while a negative value places it closer.
-   *
-   * **Initial value**: `0`
-   *
-   * | Chrome | Firefox | Safari |  Edge  |  IE   |
-   * | :----: | :-----: | :----: | :----: | :---: |
-   * | **1**  |  **1**  | **1**  | **12** | **3** |
-   *
-   * @see https://developer.mozilla.org/docs/Web/CSS/margin-top
-   * @see https://developer.mozilla.org/docs/Web/CSS/margin-bottom
-   */
-  marginY?: StandardCSSProperties['marginTop'];
-  /**
-   * The **`padding`** CSS property sets the padding on all four sides of an element. It is a shorthand for `padding-top`, `padding-right`, `padding-bottom`, and `padding-left`.
-   *
-   * | Chrome | Firefox | Safari |  Edge  |  IE   |
-   * | :----: | :-----: | :----: | :----: | :---: |
-   * | **1**  |  **1**  | **1**  | **12** | **4** |
-   *
-   * @see https://developer.mozilla.org/docs/Web/CSS/padding
-   */
-  p?: StandardCSSProperties['padding'];
-  /**
-   * The **`padding-top`** CSS property sets the height of the padding at the top of an element.
-   *
-   * **Initial value**: `0`
-   *
-   * | Chrome | Firefox | Safari |  Edge  |  IE   |
-   * | :----: | :-----: | :----: | :----: | :---: |
-   * | **1**  |  **1**  | **1**  | **12** | **4** |
-   *
-   * @see https://developer.mozilla.org/docs/Web/CSS/padding-top
-   */
-  pt?: StandardCSSProperties['paddingTop'];
-  /**
-   * The **`padding-right`** CSS property sets the width of the padding at the right side of an element.
-   *
-   * **Initial value**: `0`
-   *
-   * | Chrome | Firefox | Safari |  Edge  |  IE   |
-   * | :----: | :-----: | :----: | :----: | :---: |
-   * | **1**  |  **1**  | **1**  | **12** | **4** |
-   *
-   * @see https://developer.mozilla.org/docs/Web/CSS/padding-right
-   */
-  pr?: StandardCSSProperties['paddingRight'];
-  /**
-   * The **`padding-bottom`** CSS property sets the height of the padding on the bottom of an element.
-   *
-   * **Initial value**: `0`
-   *
-   * | Chrome | Firefox | Safari |  Edge  |  IE   |
-   * | :----: | :-----: | :----: | :----: | :---: |
-   * | **1**  |  **1**  | **1**  | **12** | **4** |
-   *
-   * @see https://developer.mozilla.org/docs/Web/CSS/padding-bottom
-   */
-  pb?: StandardCSSProperties['paddingBottom'];
-  /**
-   * The **`padding-left`** CSS property sets the width of the padding at the left side of an element.
-   *
-   * **Initial value**: `0`
-   *
-   * | Chrome | Firefox | Safari |  Edge  |  IE   |
-   * | :----: | :-----: | :----: | :----: | :---: |
-   * | **1**  |  **1**  | **1**  | **12** | **4** |
-   *
-   * @see https://developer.mozilla.org/docs/Web/CSS/padding-left
-   */
-  pl?: StandardCSSProperties['paddingLeft'];
-  /**
-   * The **`px`** property is shorthand for the CSS properties **`padding-left`** and **`padding-right`**. They set the width of the padding at the left and right side of an element.
-   *
-   * **Initial value**: `0`
-   *
-   * | Chrome | Firefox | Safari |  Edge  |  IE   |
-   * | :----: | :-----: | :----: | :----: | :---: |
-   * | **1**  |  **1**  | **1**  | **12** | **4** |
-   *
-   * @see https://developer.mozilla.org/docs/Web/CSS/padding-left
-   * @see https://developer.mozilla.org/docs/Web/CSS/padding-right
-   */
-  px?: StandardCSSProperties['paddingLeft'];
-  /**
-   * The **`paddingX`** property is shorthand for the CSS properties **`padding-left`** and **`padding-right`**. They set the width of the padding at the left and right sides of an element.
-   *
-   * **Initial value**: `0`
-   *
-   * | Chrome | Firefox | Safari |  Edge  |  IE   |
-   * | :----: | :-----: | :----: | :----: | :---: |
-   * | **1**  |  **1**  | **1**  | **12** | **4** |
-   *
-   * @see https://developer.mozilla.org/docs/Web/CSS/padding-left
-   * @see https://developer.mozilla.org/docs/Web/CSS/padding-right
-   */
-  paddingX?: StandardCSSProperties['paddingLeft'];
-  /**
-   * The **`py`** property is shorthand for the CSS properties **`padding-top`** and **`padding-bottom`**. They set the width of the padding at the top and bottom of an element.
-   *
-   * **Initial value**: `0`
-   *
-   * | Chrome | Firefox | Safari |  Edge  |  IE   |
-   * | :----: | :-----: | :----: | :----: | :---: |
-   * | **1**  |  **1**  | **1**  | **12** | **4** |
-   *
-   * @see https://developer.mozilla.org/docs/Web/CSS/padding-top
-   * @see https://developer.mozilla.org/docs/Web/CSS/padding-bottom
-   */
-  py?: StandardCSSProperties['paddingTop'];
-  /**
-   * The **`paddingY`** property is shorthand for the CSS properties **`padding-top`** and **`padding-bottom`**. They set the width of the padding at the top and bottom of an element.
-   *
-   * **Initial value**: `0`
-   *
-   * | Chrome | Firefox | Safari |  Edge  |  IE   |
-   * | :----: | :-----: | :----: | :----: | :---: |
-   * | **1**  |  **1**  | **1**  | **12** | **4** |
-   *
-   * @see https://developer.mozilla.org/docs/Web/CSS/padding-top
-   * @see https://developer.mozilla.org/docs/Web/CSS/padding-bottom
-   */
-  paddingY?: StandardCSSProperties['paddingTop'];
-
-  /**
-   * The **`typography`** property  is shorthand for the CSS properties **`font-family`**, **`font-weight`**, **`font-size`**, **`line-height`**, **`letter-spacing`** and **`text-transform``**.
-   * It takes the values defined under `theme.typography` and spreads them on the element.
-   *
-   * **Initial value**: `0`
-   *
-   * | Chrome | Firefox | Safari |  Edge  |  IE   |
-   * | :----: | :-----: | :----: | :----: | :---: |
-   * | **2**  |  **1**  | **1**  | **12** | **5.5** |
-   *
-   * @see https://developer.mozilla.org/docs/Web/CSS/font-family
-   * @see https://developer.mozilla.org/docs/Web/CSS/font-weight
-   * @see https://developer.mozilla.org/docs/Web/CSS/font-size
-   * @see https://developer.mozilla.org/docs/Web/CSS/line-height
-   * @see https://developer.mozilla.org/docs/Web/CSS/letter-spacing
-   * @see https://developer.mozilla.org/docs/Web/CSS/text-transform
-   */
-  typography?: string;
-  /**
-   * The **`displayPrint`** property sets the display value for the element when the page is printed.
-   *
-   * **Initial value**: `inline`
-   *
-   * | Chrome | Firefox | Safari |  Edge  |  IE   |
-   * | :----: | :-----: | :----: | :----: | :---: |
-   * | **1**  |  **1**  | **1**  | **12** | **4** |
-   *
-   * @see https://developer.mozilla.org/docs/Web/CSS/display
-   */
-  displayPrint?: StandardCSSProperties['display'];
-}
-
-export interface OverwriteCSSProperties {
-  /**
-   * The **`border`** CSS property is shorthand for the CSS properties **`border-width`**, **`border-style`**, and **`border-color`**. It sets an element's border.
-   *
-   * **Initial value**: `none`
-   *
-   * | Chrome | Firefox | Safari |  Edge  |  IE   |
-   * | :----: | :-----: | :----: | :----: | :---: |
-   * | **1**  |  **1**  | **1**  | **12** | **4** |
-   *
-   * @see https://developer.mozilla.org/docs/Web/CSS/border
-   */
-  border?: CSS.Property.Border | number;
-  /**
-   * The **`box-shadow`** CSS property adds shadow effects around an element's frame. You can set multiple effects separated by commas. A box shadow is described by X and Y offsets relative to the
-   * element for blur and spread radii, and by its color.
-   *
-   * **Initial value**: `none`
-   *
-   * | Chrome  | Firefox | Safari  |  Edge  |  IE   |
-   * | :-----: | :-----: | :-----: | :----: | :---: |
-   * | **10**  |  **4**  | **5.1** | **12** | **9** |
-   * | 1 _-x-_ |         | 3 _-x-_ |        |       |
-   *
-   * @see https://developer.mozilla.org/docs/Web/CSS/box-shadow
-   */
-  boxShadow?: CSS.Property.BoxShadow | number;
-  /**
-   * The **`font-weight`** CSS property specifies the weight (or boldness) of the font. The font weights available to you will depend on the `font-family` you are using. Some fonts are only
-   * available in `normal` and `bold`.
-   *
-   * **Initial value**: `normal`
-   *
-   * | Chrome | Firefox | Safari |  Edge  |  IE   |
-   * | :----: | :-----: | :----: | :----: | :---: |
-   * | **2**  |  **1**  | **1**  | **12** | **3** |
-   *
-   * @see https://developer.mozilla.org/docs/Web/CSS/font-weight
-   */
-  fontWeight?: CSS.Property.FontWeight | string;
-  /**
-   * The **`z-index`** CSS property sets the z-order of a positioned element and its descendants or flex items. Overlapping elements with a higher z-index cover those with a lower one.
-   *
-   * **Initial value**: `auto`
-   *
-   * | Chrome | Firefox | Safari |  Edge  |  IE   |
-   * | :----: | :-----: | :----: | :----: | :---: |
-   * | **1**  |  **1**  | **1**  | **12** | **4** |
-   *
-   * @see https://developer.mozilla.org/docs/Web/CSS/z-index
-   */
-  zIndex?: CSS.Property.ZIndex | string;
-}
-
-/**
- * Map of all available CSS properties (including aliases) and their raw value.
- * Only used internally to map CCS properties to input types (responsive value,
- * theme function or nested) in `SystemCssProperties`.
- */
-export interface AllSystemCSSProperties
-  extends Omit<StandardCSSProperties, keyof OverwriteCSSProperties>,
-    AliasesCSSProperties,
-    OverwriteCSSProperties {}
-
-export type SystemCssProperties<Theme extends object = {}> = {
-  [K in keyof AllSystemCSSProperties]:
-    | ResponsiveStyleValue<AllSystemCSSProperties[K]>
-    | ((theme: Theme) => ResponsiveStyleValue<AllSystemCSSProperties[K]>)
-    | SystemStyleObject<Theme>;
-};
-
-/**
- * The `SystemStyleObject` defines custom properties that will be transformed to
- * their corresponding values from the `Theme`. Other valid CSS properties are also allowed.
- */
-export type SystemStyleObject<Theme extends object = {}> =
-  | SystemCssProperties<Theme>
-  | CSSPseudoSelectorProps<Theme>
-  | CSSSelectorObject<Theme>
-  | null;
-
-export type SxProps<Theme extends object = {}> = SystemStyleObject<Theme>;
-
-// eslint-disable-next-line @typescript-eslint/naming-convention
-export function unstable_styleFunctionSx(props: object): object;
-
-// utils
-type Omit<T, K> = Pick<T, Exclude<keyof T, K>>;
+export { default as unstable_styleFunctionSx } from './styleFunctionSx';
+export * from './styleFunctionSx';
