+++ packages/material-ui-system/src/styleFunctionSx/StandardCssProperties.d.ts
@@ -1,0 +1,3 @@
+import * as CSS from 'csstype';
+
+export type StandardCSSProperties = CSS.PropertiesFallback<number | string>;
