+++ packages/material-ui-system/src/styleFunctionSx/index.d.ts
@@ -1,0 +1,6 @@
+export { default } from './styleFunctionSx';
+
+export * from './styleFunctionSx';
+export * from './AliasesCSSProperties';
+export * from './OverwriteCSSProperties';
+export * from './StandardCssProperties';
