+++ packages/material-ui-system/src/styleFunctionSx/index.js
@@ -1,0 +1,1 @@
+export { default } from './styleFunctionSx';
